feat: Update goreleaser action arguments in release.yaml
This commit updates the goreleaser action arguments in the release.yaml file.
The previously used "--rm-dist" argument has been replaced with "--clean".
This should result in a more efficient and cleaner release process.
Additionally, the environment variable for the GITHUB_TOKEN has been changed to use the default secret available in Github actions.