Removed specific Go version requirement in release workflow
I made changes to the release.yaml file to remove the specific requirement for Go version.
Previously, the workflow downloaded Go version 1.13.1 for Linux AMD64, but this is no longer necessary.
The updated workflow now uses the latest available version of Go during the build process.
This change will make the workflow more flexible and easier to modify in the future.